Move TypeTests, TypeProxyTest and SlotsTests from builtins_test.py to builtins_type_test.py
